requirements.txt
blinker==1.6.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
Flask==3.0.0
Flask-WTF==1.2.1
fonttools==4.43.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
Werkzeug==3.0.0
WTForms==3.0.1
